lint, update with tests
